refactor: move getTags to contentlayer lib
